+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Fixed `array[idx].as<JsonVariant>()` and `object[key].as<JsonVariant>()`
+* Fixed return value of `JsonObject::set()` (issue #350)
 
 v5.6.6
 ------